# internal/api/grpc/proto/auth.pb.go
// This file would normally be generated by the Protocol Buffers compiler.
// A simplified placeholder version is provided here.

package proto

// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.28.1
// 	protoc        v3.21.12
// source: api/proto/auth.proto

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

// Registration message types
type RegistrationRequest struct {
	FirstName             string `protobuf:"bytes,1,opt,name=firstName,proto3" json:"firstName,omitempty"`
	LastName              string `protobuf:"bytes,2,opt,name=lastName,proto3" json:"lastName,omitempty"`
	Nickname              string `protobuf:"bytes,3,opt,name=nickname,proto3" json:"nickname,omitempty"`
	Email                 string `protobuf:"bytes,4,opt,name=email,proto3" json:"email,omitempty"`
	AcceptedPrivacyPolicy bool   `protobuf:"varint,5,opt,name=acceptedPrivacyPolicy,proto3" json:"acceptedPrivacyPolicy,omitempty"`
}

type RegistrationResponse struct {
	RegistrationSessionId string `protobuf:"bytes,1,opt,name=registrationSessionId,proto3" json:"registrationSessionId,omitempty"`
	CodeExpires           int64  `protobuf:"varint,2,opt,name=codeExpires,proto3" json:"codeExpires,omitempty"`
	Code                  string `protobuf:"bytes,3,opt,name=code,proto3" json:"code,omitempty"`
}

type ConfirmEmailRequest struct {
	RegistrationSessionId string `protobuf:"bytes,1,opt,name=registrationSessionId,proto3" json:"registrationSessionId,omitempty"`
	Code                  string `protobuf:"bytes,2,opt,name=code,proto3" json:"code,omitempty"`
}

type EmptyResponse struct{}

type ResendCodeEmailRequest struct {
	RegistrationSessionId string `protobuf:"bytes,1,opt,name=registrationSessionId,proto3" json:"registrationSessionId,omitempty"`
}

// Login message types
type SendLoginCodeEmailRequest struct {
	Email string `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
}

type CodeResponse struct {
	CodeExpires int64  `protobuf:"varint,1,opt,name=codeExpires,proto3" json:"codeExpires,omitempty"`
	Code        string `protobuf:"bytes,2,opt,name=code,proto3" json:"code,omitempty"`
}

type ConfirmLoginRequest struct {
	Email string `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
	Code  string `protobuf:"bytes,2,opt,name=code,proto3" json:"code,omitempty"`
}

// Token message types
type TokenResponse struct {
	AccessToken  string `protobuf:"bytes,1,opt,name=accessToken,proto3" json:"accessToken,omitempty"`
	RefreshToken string `protobuf:"bytes,2,opt,name=refreshToken,proto3" json:"refreshToken,omitempty"`
}

type RefreshTokenRequest struct {
	RefreshToken string `protobuf:"bytes,1,opt,name=refreshToken,proto3" json:"refreshToken,omitempty"`
}

type ValidateTokenRequest struct {
	AccessToken string `protobuf:"bytes,1,opt,name=accessToken,proto3" json:"accessToken,omitempty"`
}

type UserResponse struct {
	UserId   uint64 `protobuf:"varint,1,opt,name=userId,proto3" json:"userId,omitempty"`
	Nickname string `protobuf:"bytes,2,opt,name=nickname,proto3" json:"nickname,omitempty"`
	Email    string `protobuf:"bytes,3,opt,name=email,proto3" json:"email,omitempty"`
	Role     string `protobuf:"bytes,4,opt,name=role,proto3" json:"role,omitempty"`
}